calc logps p5866

+++ calc_logps_sims.py
@@ -60,7 +60,7 @@ def load_inference_data():
 
 
 def get_constants(vlps):
-    k = gpl.k.astype('float32')
+    k = gpl.k
     vref = vlps[0].astype('float32')
 
     return k, vref
